requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
sqlalchemy==2.0.29
python-dotenv==1.0.1
passlib==1.7.4
psycopg2-binary==2.9.9
pydantic==2.7.1 
pydantic-core==2.18.2
Jinja2==3.1.4 
MarkupSafe==2.1.5 
aiosmtplib==2.0.2 
blinker==1.8.2 
dnspython==2.6.1 
email-validator==2.1.1 
fastapi-mail==1.4.1 
pydantic_settings==2.2.1
ecdsa==0.19.0
pyasn1==0.6.0 
python-jose==3.3.0 
rsa==4.9 
six==1.16.0